refactor: mutate fecher 폴더 구조 변경
