+++ src/apis/fetchers/category/delete-category/fetcher.ts
