+++ src/apis/fetchers/document/create-ai-pick/fetcher.ts
