+++ src/apis/fetchers/document/delete-document/fetcher.ts
