+++ src/apis/fetchers/quiz/delete-quiz/fetcher.ts
