+++ src/app/(routes)/(document)/(detail)/document/[documentId]/components/ai-pick.tsx
@@ -11,10 +11,10 @@ import { cn } from '@/lib/utils'
 import { motion, AnimatePresence } from 'framer-motion'
 import { SwitchCase } from '@/components/react/switch-case'
 import { useMutation, useQueryClient } from '@tanstack/react-query'
-import { createAiPick } from '@/apis/fetchers/document/create-ai-pick'
+import { createAiPick } from '@/apis/fetchers/document/create-ai-pick/fetcher'
 import { useSession } from 'next-auth/react'
 import { useParams } from 'next/navigation'
-import { toggleBookmark } from '@/apis/fetchers/key-point/toggle-bookmark'
+import { toggleBookmark } from '@/apis/fetchers/key-point/toggle-bookmark/fetcher'
 import { DocumentStatus } from '@/apis/types/dto/document.dto'
 import { useToast } from '@/hooks/use-toast'
 import { GeneratingPicks } from '../ui/generating-picks'
