+++ src/app/(routes)/(document)/create/page.tsx
@@ -7,7 +7,7 @@ import dynamic from 'next/dynamic'
 import { CreateDocumentProvider } from './contexts/create-document-context'
 import { Header } from './components/header'
 import { TitleInput } from './components/title-input'
-import { createDocument } from '@/apis/fetchers/document/create-document'
+import { createDocument } from '@/apis/fetchers/document/create-document/fetcher'
 import { useRouter } from 'next/navigation'
 import { useState } from 'react'
 import { useToast } from '@/hooks/use-toast'
