+++ src/app/(routes)/(document)/modify/[documentId]/page.tsx
@@ -8,7 +8,7 @@ import { Header } from '../components/header'
 import Loading from '@/components/loading'
 import { TitleInput } from '../components/title-input'
 import VisualEditor from '../components/visual-editor'
-import { updateDocumentContent } from '@/apis/fetchers/document/update-document-content'
+import { updateDocumentContent } from '@/apis/fetchers/document/update-document-centent/fetcher'
 import { useState } from 'react'
 import { useToast } from '@/hooks/use-toast'
 import { ToastAction } from '@/components/ui/toast'
