+++ src/app/(routes)/(tabs)/main/components/make-quiz-drawer-dialog.tsx
@@ -10,7 +10,7 @@ import { Checkbox } from '@/components/ui/checkbox'
 import { CategoryDTO } from '@/apis/types/dto/category.dto'
 import { useCheckList } from '@/hooks/use-check-list'
 import { useMutation } from '@tanstack/react-query'
-import { createQuizzes } from '@/apis/fetchers/quiz/create-quizzes'
+import { createQuizzes } from '@/apis/fetchers/quiz/create-quizzes/fetcher'
 import { useSession } from 'next-auth/react'
 import { useRouter } from 'next/navigation'
 import Loading from '@/components/loading'
