+++ src/app/(routes)/(tabs)/profile/notification/email/components/email-setting-form.tsx
@@ -1,8 +1,8 @@
 'use client'
 
 import { API_ENDPOINT } from '@/apis/api-endpoint'
-import { verifyEmail } from '@/apis/fetchers/auth/verify-email'
-import { verifyEmailCheck } from '@/apis/fetchers/auth/verify-email-check'
+import { verifyEmail } from '@/apis/fetchers/auth/verify-email/fetcher'
+import { verifyEmailCheck } from '@/apis/fetchers/auth/verify-email-check/fetcher'
 import { Button } from '@/components/ui/button'
 import { useToast } from '@/hooks/use-toast'
 import { actionRevalidatePath } from '@/lib/revalidate'
