+++ src/app/(routes)/(tabs)/repository/components/category-list.tsx
@@ -19,7 +19,7 @@ import { useMutation, useQueryClient } from '@tanstack/react-query'
 import { Category } from '@/apis/fetchers/category/get-categories/fetcher'
 import { useSession } from 'next-auth/react'
 import icons from '@/constants/icons'
-import { reorderCategory } from '@/apis/fetchers/category/reorder-category'
+import { reorderCategory } from '@/apis/fetchers/category/reorder-category/fetcher'
 import { cn } from '@/lib/utils'
 import CreateCategoryDialog from '@/components/create-category-dialog'
 import Loading from '@/components/loading'
