+++ src/app/(routes)/(tabs)/review/picks/components/search-result.tsx
@@ -1,7 +1,7 @@
-import { SearchKeyPointsResponse } from '@/apis/fetchers/key-point/search-key-points'
+import { SearchKeyPointsResponse } from '@/apis/fetchers/key-point/search-key-points/fetcher'
 import { CommonLayout } from '@/components/common-layout'
 import { useMutation, useQueryClient } from '@tanstack/react-query'
-import { toggleBookmark } from '@/apis/fetchers/key-point/toggle-bookmark'
+import { toggleBookmark } from '@/apis/fetchers/key-point/toggle-bookmark/fetcher'
 import { useSession } from 'next-auth/react'
 import { SearchForm } from '@/components/search-form'
 import { KeyPointCard } from './key-point-card'
