+++ src/app/(routes)/(tabs)/review/picks/page.tsx
@@ -4,11 +4,11 @@ import { CommonLayout } from '@/components/common-layout'
 import { GetBookmarksResponse } from '@/apis/fetchers/key-point/get-bookmarks/fetcher'
 import { useSession } from 'next-auth/react'
 import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query'
-import { toggleBookmark } from '@/apis/fetchers/key-point/toggle-bookmark'
+import { toggleBookmark } from '@/apis/fetchers/key-point/toggle-bookmark/fetcher'
 import Loading from '@/components/loading'
 import { NoPicks } from './components/no-picks'
 import { usePathname, useRouter, useSearchParams } from 'next/navigation'
-import { searchKeyPoints } from '@/apis/fetchers/key-point/search-key-points'
+import { searchKeyPoints } from '@/apis/fetchers/key-point/search-key-points/fetcher'
 import { SearchResult } from './components/search-result'
 import { KeyPointCard } from './components/key-point-card'
 import { useGetBookmarksQuery } from '@/apis/fetchers/key-point/get-bookmarks/query'
