+++ src/app/api/auth/[...nextauth]/auth.ts
@@ -1,5 +1,5 @@
 import NextAuth, { Account, DefaultSession, NextAuthResult } from 'next-auth'
-import { signIn as signInAPI } from '@/apis/fetchers/auth/sign-in'
+import { signIn as signInAPI } from '@/apis/fetchers/auth/sign-in/fetcher'
 import Kakao from 'next-auth/providers/kakao'
 import Google from 'next-auth/providers/google'
 import { getUser } from '@/apis/fetchers/user/get-user/fetcher'
